requirements.txt
torch
transformers
accelerate
bitsandbytes
psutil
pynvml
pandas
matplotlib
tqdm
pyyaml
